Fixed bug in loadProgram() function
Fixed a bug in the loadProgram() function.  For some reason using
prog.size() in the for-loop was not working correctly and
resulted in breaking the loop early.  Instead, I just used a
variable to store the value and used that in the loop.

There is still an issue with Assertion Failure in debug mode,
however the program runs fine while in Release mode.  I will
need to research why this happens.