Fixed the Assertion Failed bug
Fixed the issue with the Assertion Failed vector subscript out
of range bug.  This was happening because when a StackVM object
was initialized, the memory vector used reserve() when allocating
memory.  This does not affect size().  So when a for-loop tries
to write information, it is considered out of bounds.  I am now
using resize() instead.