Wrote main code for MyVM
Wrote code for MyVM.  The goal of this progam will be to control
the two pieces of the virtual machine.  The code for this is not
yet complete.  I still want to use somthing other than the call
to system() for executing the other exe files.  Will need to
research this.